fix: 修复notifier延时,以及shm_conf数据类型和注释问题

+++ transport/shm/state.h
@@ -44,7 +44,7 @@ private:
     std::atomic<bool> need_remap_ = {false};            // 
     std::atomic<uint32_t> seq_ = { 0 };                 // 
     std::atomic<uint32_t> reference_count_ = { 0 };     // 引用计数
-    std::atomic<uint32_t> ceiling_msg_size_;            // 
+    std::atomic<uint64_t> ceiling_msg_size_;            // 
 };
 
 } // transport
